-Change time to float, its in seconds so need decimals and all kinds of new mathy shenanigans -Check if we have animation before adding active checkbox

+++ src/com/blastedstudios/gdxworld/plugin/mode/animation/AnimationEditor.java
@@ -83,7 +83,7 @@ public class AnimationEditor extends Table{
 			animation.setAnimations(newStructs);
 			animation.setName(nameField.getText());
 			animation.setRepeat(repeatCheckbox.isChecked());
-			animation.setTotalTime(Long.parseLong(totalTimeField.getText()));
+			animation.setTotalTime(Float.parseFloat(totalTimeField.getText()));
 		}
 		return animation;
 	}
